refactor: split quest card
